requirements.txt
# Core dependencies
python-dotenv>=1.0.0
colorama>=0.4.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Statistical analysis
scipy>=1.10.0

# Visualization (optional, for analysis tools)
matplotlib>=3.7.0
seaborn>=0.12.0

# AI integration (optional)
openai>=0.27.0

# Database
sqlite3  # Usually comes with Python

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0